refactor code : organize classes. remove require main from rspec file. use let to instanciate class instances.
