+++ spec/tic_tac_toe_spec.rb
@@ -1,16 +1,20 @@
-require('./bin/main')
-require('./lib/tic_tac_toe')
+# require_relative '../bin/main'
+require_relative '../lib/tic_tac_toe'
+require_relative '../lib/player'
+require_relative '../lib/game'
 
-board = Board.new
+describe Game do
+  let(:game) { Game.new }
 
-describe '#welcome_player' do
-  it 'welcomes player' do
-    expect(welcome_player('A', 'X')).to eq("\nA will play with X\n\n")
+  describe '#welcome_player' do
+    it 'welcomes player' do
+      expect(game.welcome_player('A', 'X')).to eq("\nA will play with X\n\n")
+    end
   end
-end
-
-describe '#display_board' do
-  it 'displays board' do
-    expect(display_board(board.board)).to eq("+---+---+---+\n|   |   |   |\n+---+---+---+\n|   |   |   |\n+---+---+---+\n|   |   |   |\n+---+---+---+\n\n")
+  let(:board) { Board.new }
+  describe '#display_board' do
+    it 'displays board' do
+      expect(game.display_board(board.board)).to eq("+---+---+---+\n|   |   |   |\n+---+---+---+\n|   |   |   |\n+---+---+---+\n|   |   |   |\n+---+---+---+\n\n")
+    end
   end
 end
